test: move offline playlist check into test suite
