+++ tests/commands/playlist/test.test.ts
@@ -1,4 +1,5 @@
-import child_process from 'node:child_process'
+import child_process, { execSync } from 'node:child_process'
+import path from 'node:path'
 import { pathToFileURL } from 'node:url'
 import { promisify } from 'node:util'
 import * as fs from 'fs-extra'
@@ -9,6 +10,7 @@ const exec = promisify(child_process.exec)
 type ExecError = {
   status: number
   stdout: string
+  stderr: string
 }
 
 const ENV_VAR = 'cross-env DATA_DIR=tests/__data__/input/data ROOT_DIR=tests/__data__/output'
@@ -53,6 +55,20 @@ describe('playlist:test', () => {
           content(`tests/__data__/expected/playlist_test/${filepath}`)
         )
       })
+    }
+  })
+
+  it('fails when the network is unavailable', () => {
+    const preload = path.resolve('tests/__data__/input/offline_dns.cjs').replaceAll('\\', '/')
+    const cmd =
+      `cross-env NODE_OPTIONS="--require=${preload}" npm run playlist:test -- streams/af.m3u`
+
+    try {
+      execSync(cmd, { encoding: 'utf8' })
+      throw new Error('Expected playlist:test to fail offline')
+    } catch (error) {
+      const output = `${(error as ExecError).stdout || ''}${(error as ExecError).stderr || ''}`
+      expect(output).toContain('Internet connection is required for the script to work')
     }
   })
 })
